add reverse_min_collatz

+++ cs4414rust_tutor/collatz/main.rs
@@ -7,7 +7,7 @@ fn main() {
   }
   // below is an alternative
   //let temp: Vec<String> = env::args().collect();
-  let i:u32 = match (env::args().collect::<Vec<String>>())[1].trim().parse() {
+  let i:u64 = match (env::args().collect::<Vec<String>>())[1].trim().parse() {
     Ok(num) => num,
     Err(_) => return
   };
@@ -15,9 +15,12 @@ fn main() {
 }
 
 #[allow(non_snake_case)]
-fn collatz(N: u32) -> u32 {
+fn collatz(N: u64) -> u64 {
   if N == 1 { return 0; }
   match N % 2 {
+    // an expression, which is not followed by punctuation like semicolon, comma
+    // at the the end of a function is treated as a return value, note here
+    // no commas for each arm of the match
     0 => { 1 + collatz(N/2) }
     _ => { 1 + collatz(N*3+1) }
   }
